elfkeep.dmm Map for Ratkeep!
This map is designed to be used for very short or very long rounds. Featuring dungeons, exploration and content geared towards three types of players, that being Roleplayers, PvE Adventurers, and even the capacity for a King of the Hill-Type RDM Deathmatch.

I highly reccomend giving it a try and telling me if you can see anything that needs work, such as lighting or other mapping issues. I plan on expanding the map to a much larger one in the future.