Elfkeep Adventure Map
Adds elfkeep as a selectable map. Elfkeep works for short chaos rounds, PvE Co-Op, Tutorials and RP-Sessions. 

elfkeep is a customized version of roguetown with an Wood-Elf Theme, designed for lots of Thief/Archer Gameplay and turning the formula on it's heels to promote fun narrative chaos. 

Be aware; adventurers spawn in the keep itself and may kill the king immediately, though intended to work together protecting the realm.

Be aware; the Priest spawns sorrounding by druids in the woods. 

Be aware; Vampires live in the church attic. (They are friendly?)

Be aware; The Wizard's Tower is now an Academy.

Be aware; The Mercenaries guild is now a Monster Hunter's Guild.

Be aware; The Merchant's Guild is now a Free Trading Post run by nobles and shopkeepers.

Be aware; the city is built on an ancient dungeon that nobody has truly found the bottom of. (T.B.A)

Elfkeep is designed for H-RP and M-RP players in mind as something to change up the round pacing; as well as a map option to inspire. Feel free to edit/change the map as you like and use it as inspiration for your own works!